refact: router에서 쿠키 검증 및 메인라우터 통합
