+++ src/route/PrivateRouter.tsx
@@ -1,12 +1,16 @@
 // import React from 'react'
 import { Outlet, Navigate } from 'react-router-dom'
-// import { getUserInfo } from '../app/slices/user/userSlice'
-// import { useAppSelect } from '../app/configureStore.hooks'
+import { useAppSelect } from '../redux/store.hooks'
 
-const PrivateRoute = () => {
+const AuthRoute = () => {
   // const user = useAppSelect(getUserInfo) // Validation logic in here
-  const user = 'a'
-  return user ? <Outlet /> : <Navigate to="/login" />
+  const auth = useAppSelect(select => select.auth.isAuthorized)
+  return auth ? <Outlet /> : <Navigate to="/login" />
 }
 
-export default PrivateRoute
+const UnauthRoute = () => {
+  const auth = useAppSelect(select => select.auth.isAuthorized)
+  return !auth ? <Outlet /> : <Navigate to="/" />
+}
+
+export { AuthRoute, UnauthRoute }
